+++ NEWS.md
@@ -65,6 +65,7 @@ Active logins are now necessary to use webservices such as CSI:FingerID, see e.g
 - `plotScores()`
     - split bars for sets
     - only split bars if results are present for >1 sets and/or consensus algorithms
+- Fixed: `selectIons()` didn't properly handle empty components objects
 
 
 # patRoon 2.1